Implement a more ergonomic API, add docs
Implementation details are hidden and there is no more magic macro
creating and expanding c functions. It is hidden through a feature flag
that is enabled by default.

Now, if a user wants to extend the API with their own stuff and make a
library on top of zsh-module, they can be sure it will not conflict.

Just disable the default features.

Also, add some docs that help me keep track of what stuff does